misc: moved `input.hpp` into `tui.hpp`, input namespace; updated all the examples to use this version of `tui.hpp`; new basic example
